LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/manual-scroll-push-expected.txt


FAIL scroll: scroll() should work on a push navigation assert_equals: expected 0 but got 431
